Add notes for when to override methods on `Metric` and delineate the ones that should vs. should not be overriden (#1660)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1660

This is done in preparation for a refactor; I was reminding myself exactly how metrics work and thought it would be useful to add in-code guidance for how to implement one's own metrics. This also will help clarify what is going on in this refactor.

NOTE: We could even consider adding a base `MetricTest` that validates that only the methods that we expect to see overriden, are overriden?

Reviewed By: bernardbeckerman

Differential Revision: D46455686

fbshipit-source-id: 2de5cba9c9ae4c76061c54846c94a13e3fd46a87