Model fit metrics for logging (#1682)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1682

This commit adds metrics in order to quantify and log model fit quality for each *experimental* metric. To start, the commit adds metrics based on the posterior statistics of the model, which can be extended readily by adding to the `fit_metrics` dict, and can be generalized with other metric types in follow up work.

Reviewed By: Balandat

Differential Revision: D46816506

fbshipit-source-id: 0e4f9d9d8f4030b9793bdcf9ec5c218fccb91990