Store floats exactly in Pandas DataFrames; require exact equality for Pandas DataFrames to be considered equal (#1750)
Summary:
Pandas `to_json` by default stores 10 decimal places of floats. This leads to precision loss. The fix is not completely trivial, because storing many more decimal places would not guarantee recoverability of very small floats. A better way to store floats exactly as strings is to use `repr`, which will store the float in as few digits as possible while maintaining that `float(repr(value)) = value`.

* Use `repr` to encode floats as strings without precision loss when storing DataFrames as json
* Store dtype info so that data can be recovered exactly. (Otherwise, there would be no way of knowing that "3.1234" should be a float but "3_1" is an int, sine 3.1234 and 3_1 are both valid numbers in Python).
* Add DataFrame encoders and decoders
* Require DataFrames to be exactly equal by default for equality checks to pass

A downside of this change is that it stores the format in which DataFrames are encoded, so now the DataFrame decoder must support both the old and new formats.

Pull Request resolved: https://github.com/facebook/Ax/pull/1750

Test Plan:
* Added a unit test that reproduces the behavior and ensures that DataFrames stored in both the old and new formats are properly decoded.
* Added test cases for exact vs. approximate DataFrame equality.

Differential Revision: D47922407

Pulled By: esantorella

fbshipit-source-id: 772dd6d30727cb3093a4826fb79fb38bb7ee00b8