Allow customizing the name of a GenerationStep (#1754)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1754

Currently, the generation steps carry the names of the `ModelSpec`s they are constructed from, without a way of customizing it. This makes it so that every GStep that is constructed from `BoTorch` (MBM) registry entry produces the name `BoTorch`. This is not ideal since MBM is highly customizable through `model_kwargs`. Having a way of re-naming customized models without having to define a new `ModelSpec` would simplify things.

Reviewed By: lena-kashtelyan

Differential Revision: D47747138

fbshipit-source-id: 4ee8cb49dc8e8d13876880faa8d73517b70c8313