Refactor Curve & Tensorboard metrics (#1763)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1763

This diff leverages recent changes to metrics in D46484953 that get rid of the nasty `classmethod`s that were throughout the code. This makes it much less clunky to have minor configurable variants of already existing metrics without having to introduce a bunch of subclasses.

A follow-up diff will deal with refactoring / improving the actual logic of reading tensorboard logs. Currently, things are rather inefficient in that we may read a lot of logs that we don't actually need, and we're using a custom and legacy way of reading the logs.

Reviewed By: bernardbeckerman

Differential Revision: D46385253

fbshipit-source-id: bb7efc3bfc644c690643b8bdebf476de772452ad