Migrate generation strategy logic to generation step 1/n (#1758)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1758

This diff is the first in a stack of diffs that will migrate much of the logic from generation strategy into generation step (and then subsquently generation node).

In this diff we move `num_trials_to_gen_and_complete_in_curr_step()` to generation step. As a part of this we also move a few helper functions:
- trial_indices_by_step
- num_can_complete_this_step
- num_completed_this_step

There is a bit of weirdness with how the experiment is being set right now that i've left a todo to clean up as we move through this restructuring

In follow up diffs plan to work my way up to the gen_multiple function call moving out pieces as we go

Reviewed By: lena-kashtelyan

Differential Revision: D47956491

fbshipit-source-id: 6d1d39aeb450b5b7754eac55df40179d661747f8