Fix sebo tutorial (#1801)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1801

Sebo tutorial was failing (marked "flaky" because it has never succeeded since creation) because it was not included in the target graph. Also,  mixed use of options.get and options.pop was causing issues with input constructor validations. Both issues are resolved now.

Finally, I changed the number of batches from 40 to 20. Let me know if this is acceptable -- it puts our runtime at 8:46 in opt mode which is still close to timeout but should be consistently safe.

Reviewed By: esantorella

Differential Revision: D48737511

fbshipit-source-id: 7db693468890eba332801baa4228fd44e0bc2d73