Prevent duplicate trial_index+arm_name identifiers in report utils (#1807)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1807

See OSS task for motivation.
"When combining trial index and arm name to one identifier there's currently the possibility that the resulting identifier is not unique. For example, trial index 17 and arm 7_0 will yield 177_0, but trial index 1 and arm 77_0 will also yield 177_0.

This will in line 463 in report_utils.py where duplicate entries are searched for, lead to a data frame without results. An easy fix is to just add a hyphen between trial index and arm name when creating the ID string."

- OSS user wlanghans

Reviewed By: bernardbeckerman

Differential Revision: D48829031

fbshipit-source-id: 3b1a366e775696557d3f409b7fcddefa8879df6b