Simplify and migrate maybe_move_to_next step from GenStrat to GenStep 4/n (#1806)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1806

This is a step in the direction of migrating much of the logic from generation strategy to generation step, and then to generation node

This diff significantly simplifies the maybe_move_to_next_step logic on GenerationStrategy by creating the maybe_step_completed method on GenerationStep. We also add a couple of headers to try to start giving some structure to the GenerationStep class as it grows larger.

We are slowly working our way to functions "higher in the generation strategy tree" pulling out those that are relevant only for generation steps.

(potentially many) More diffs to follow

Reviewed By: lena-kashtelyan

Differential Revision: D48755965

fbshipit-source-id: 3575d490d90f5781913ede71027303cb465085dc