Temporal Kernel
Summary:
see title. This is useful for modeling time of day.

This is a kernel implementation of the kernel. A bit more work would be needed to automated generating the input arguments, without manually specifying them via `covar_module_options`. This is convenient for now, as it unblocks its use, while leaving open options for how to treat temporal features more generally in Ax.

Reviewed By: Balandat

Differential Revision: D47633553