Temporal Kernel (#1831)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1831

see title. This is useful for modeling time of day.

This is a kernel implementation of the kernel. A bit more work would be needed to automated generating the input arguments, without manually specifying them via `covar_module_options`. This is convenient for now, as it unblocks its use, while leaving open options for how to treat temporal features more generally in Ax.

Reviewed By: Balandat

Differential Revision: D47633553

fbshipit-source-id: c534e4b0dbca2c14515dce3bd56c89fbf351edfe