Pass in data_type to data_and_evaluations_from_raw_data (#1841)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1841

`data_and_evaluations_from_raw_data` infers the DataType from the format of the `raw_data`. As a result, it was possible to attach the wrong data with the wrong type to the experiment, which leads to errors down the line (e.g., in SQA storage or while updating the trial data).

This passes in the data_type from trial to `data_and_evaluations_from_raw_data`; checks that the format of the `raw_data` is compatible with the data dtype; and raises an informative error otherwise.

Differential Revision: D49174012

fbshipit-source-id: 197ca5ebbb1647657d8774917d0237db438c8838