Migrate most of the current_generator_run_limit() logic to GenStep 6/n (#1849)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1849

This moves much of the logic for current_generator_run_limit() to generation step.

This is a step in the direction of migrating much of the logic from generation strategy to generation step, and then to generation node.

There should only be a few more methods from GenerationStrategy that will migrate to generationstep or node

Reviewed By: lena-kashtelyan

Differential Revision: D49301526

fbshipit-source-id: 901b8464a8deb4733b7e220bc44b5f5d74229eda