Give generation nodes a required name (#1855)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1855

This diff adds a required name property in the GenerationNode class. This is important for (1) storage, most important and (2) being able to surface relevant errors to users that point to a specific node causing the error.

From a big picture, this is important for the GenerationStep -> GenerationNode migration

also thank lena-kashtelyan for doing the majority of this diff after a discussion re: the field unique_id failing tests for GenNodes

Differential Revision: https://internalfb.com/D49424161

fbshipit-source-id: d57fba08eb29ef3d3d78d4dd2e2532ada6de878e