Moves up the get_data_for_fit method and use_update property from Step to Node 3/n (#1856)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1856

This is the diff is in the set of diffs that will work to migrate much of the logic from GenerationStep to GenerationNode. This diff finishes moving both model selection methods from step to node.

In this diff we move up the use_update field and the get_data_for_fit function In follow up diffs we'll:
(1) move up completion criteria to generation Node
(2) move the dedup logic into model spec
(3) pull the remaining model related methods from genstrat directly to gennode
(4) move up the trial related methods

Reviewed By: lena-kashtelyan

Differential Revision: D49305797

fbshipit-source-id: 3d22ce74596b3e2f9ec340b5613032042c445d61