Stop ignoring unused kwargs in MAP SAAS (#1863)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1863

Change `fit` method of legacy `BotorchModel` so that `prior` is only passed to `model_constructor` if the prior is not `None`, because not all models accept a prior. This is somewhat ugly, with the benefit that model constructors are no longer forced to accept the `prior` kwarg if they don't use it.

Reviewed By: SebastianAment

Differential Revision: D49501978

fbshipit-source-id: d1afb1bee252189cb9fcb6b9a0d41c9822fbbdcd