Move GPEI and MOO to MBM setups (#1865)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1865

Replace GPEI and MOO with Modular Botorch defaults. Under the hood this will select the appropriate GP and either EI variant (noisy, hypervolume, etc) based on the shape of the data coming in. This generally streamlines things, and getting all our models on MBM will be a significant reduction in our team's maintenance burden.

Reviewed By: esantorella

Differential Revision: D49064726

fbshipit-source-id: 9390f7a5bb16fc3b3808bad9fabfb368ae15b110