Enable setting labels in the batch comparison plot (#1881)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1881

A few weeks ago when the batch comparison plot was being shown to some collaborators outside the team, there's was some brief confusion around what the axes were since the "Batch XX" labels are not always super descriptive. It'd be helpful to be able to give more meaningful labels for future times that these plots are shared.

Reviewed By: Balandat

Differential Revision: D49746704

fbshipit-source-id: f043ef97fc64824640140f9ac1154e87a467010d