Run fewer trials for slow benchmark problems
Summary: Some benchmark problems are very slow. This can be due either to the runner (as with MNIST) or to the method (as with fully Bayesian methods). In many cases, running trials can actually enhance our ability to compare methods against each other, because all methods perform well on a problem after 100 trials, but methods vary a lot in their performance after 20-30 trials.

Differential Revision: D49852782