Run fewer trials for slow benchmark problems (#1888)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1888

Some benchmark problems are very slow. This can be due either to the runner (as with MNIST) or to the method (as with fully Bayesian methods). In many cases, running trials can actually enhance our ability to compare methods against each other, because all methods perform well on a problem after 100 trials, but methods vary a lot in their performance after 20-30 trials.

Reviewed By: Balandat

Differential Revision: D49852782

fbshipit-source-id: d109e34135632c83127ff5b0cae6870abca7aba8