Make target_value required for task parameters
Summary:
Task parameters have a "target_value" attribute, but we haven't been using it; it has previously only been used for fidelity parameters. But at the core level, there's nothing in its implementation that would restrict its use to fidelity parameters. But all of our applications of multi-task models have a target task, and the new MultiTaskDataset in botorch requires specifying a target task.

In order to migrate our multi-task models to use MultiTaskDataset (D49604249), we need to have a target task. This diff makes target_value a required attribute for task parameters, the same way it is currently required for fidelity parameters.

Propagating the target value to the search space digest will happen in a stacked diff.

Transforms that create task parameters (TrialAsTask and MetricsAsTask) are adjusted to appropriately specify the target value.

Differential Revision: D49659731

fbshipit-source-id: 47d66c06f86f58722703698d02d4626a38f46967