Add target task values to SearchSpaceDigest
Summary: Earlier in this stack we add "target_value" as a required attribute for task parameters, like it has been for fidelity parameters. The first place in the stack where "target_value" is considered to be exclusively for fidelity parameters is in SearchSpaceDigest, so this adjusts the logic there to also track target values for task parameters. And renames the attribute, to make it more clear that it is not restricted to fidelity parameters.

Differential Revision: D49701286

fbshipit-source-id: ca3b7af107f371d9f501a3fe0b90fec00a85d726