Make target_value required for task parameters (#1893)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1893

Task parameters have a "target_value" attribute, but we haven't been using it; it has previously only been used for fidelity parameters. But at the core level, there's nothing in its implementation that would restrict its use to fidelity parameters. But all of our applications of multi-task models have a target task, and the new MultiTaskDataset in botorch requires specifying a target task.

In order to migrate our multi-task models to use MultiTaskDataset (D49604249), we need to have a target task. This diff makes target_value a required attribute for task parameters, the same way it is currently required for fidelity parameters.

Propagating the target value to the search space digest will happen in a stacked diff.

Transforms that create task parameters (TrialAsTask and MetricsAsTask) are adjusted to appropriately specify the target value. For MetricsAsTask the correct target value is clear. For TrialAsTask it may not be (should it be trial 0 if that is an LRT? Should it be the last trial if it is a fully sequential experiment?). Moreover, the user may want to use the model to predict multiple tasks (e.g. to do cross validation on each task in the model). Fortunately, choosing the "right" target value is not essential at this point. Currently with TrialAsTask, we require specifying the target task as part of the prediction in every call to predict or gen. That will continue to work exactly as-is. So the way we currently use TrialAsTask, the behavior will not be affected by which trial is chosen as the target task. In the future, we will be able to replace specifying the target task as a fixed feature to `gen` with specifying the target task as a config option to the TrialAsTask transform. We can make that transition later.

Reviewed By: saitcakmak

Differential Revision: D49659731

fbshipit-source-id: cf1c6c8ba842e9006bc317b9783336d883597dbe