Add target task values to SearchSpaceDigest (#1894)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1894

Earlier in this stack we add "target_value" as a required attribute for task parameters, like it has been for fidelity parameters. The first place in the stack where "target_value" is considered to be exclusively for fidelity parameters is in SearchSpaceDigest, so this adjusts the logic there to also track target values for task parameters. And renames the attribute, to make it more clear that it is not restricted to fidelity parameters.

Reviewed By: saitcakmak

Differential Revision: D49701286

fbshipit-source-id: c3c1dce59e41a27991a0fe858aad3033093eda37