Deprecate `use_update` as discussed (#1895)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1895

The main use case for `use_update` in generation strategy is a stateful model, where we really don't want to restore the state on each call to `fit` and instead want to expect a call to `update` with only the new data. I think at this point supporting `use_update` just adds lots of complexity to generation strategy, and it'll be great to see that complexity gone (this diff shows it in the deleted code).

Reviewed By: saitcakmak

Differential Revision: D49465655

fbshipit-source-id: 0352055c9997e524b3b4f3da1aee78cd90b70126