Update typing of TEvaluationFunction
Summary:
I came across this erroneous type while working on something unrelated. This diff trades of a bunch of pyre-fixmes for a couple new ones, and updates the type hint with what I believe is the correct one.

An alternative type that doesn't require any pyre-fixme/ignores is `Callable[[TParameterization, ...], TEvaluationOutcome]` but I don't think it is the correct one. `...` denotes that there can be any number of `TParameterization` arguments, which `Optional[float]` is not.

Differential Revision: D50352771