Update typing of TEvaluationFunction (#1911)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1911

I came across this erroneous type while working on something unrelated. This diff trades of a bunch of pyre-fixmes for a couple new ones, and updates the type hint with what I believe is the correct one.

An alternative type that doesn't require any pyre-fixme/ignores is `Callable[[TParameterization, ...], TEvaluationOutcome]` but I don't think it is the correct one. `...` denotes that there can be any number of `TParameterization` arguments, which `Optional[float]` is not.

Reviewed By: mpolson64

Differential Revision: D50352771

fbshipit-source-id: 76b604cddf7cf1cba9e6fd762f398c6733e883b6