Add generation_node_name arg to BaseTrial (#1919)
Summary:

As we transition to using GenerationNodes by default in GenerationStrategy, we want to add in a geneartion_node_name argument to Trial Objects. This arg is analogous to the generation_step_index, but more flexibile.

For backwards compatibility purposes, we aren't replacing generation_step_index at this time. I anticipate, over time we will be able to do so.


Things in the pipeline:
(1) Update other code impacted by this trial object run chage
(2) Add the transition_to field to transition criterion class
(3) Update the transition criterion class to check on a per node basis, instead of per experiment
(4) add is_complete to generationNode and then use that in generation Strategy for moving forward
(5) [Mby] skip max trial criterion addition if numtrials == -1
(6) add transition criterion to the repr string + some of the other fields that havent made it yet

Reviewed By: lena-kashtelyan

Differential Revision: D50279168