Fix TransitionCriterion to be based on Trials from a specific node instead of entire experiment
Summary:
During the dev process we noticed that the current implementation of completion criteria relied on full experiment trials, instead we only want to pass transition criteria for Max and Min based on trials that were generated from the current node. This updates those criterion is_met methods to follow this logic.

We also add unit tests and repr method for printing the classes, this helps with debugging

Things in the pipeline:
(1) Use the transition criterion to determine if a node is complete
(2) add is_complete to generationNode and then use that in generation Strategy for moving forward
(3) [Mby] skip max trial criterion addition if numtrials == -1
(4) add transition criterion to the repr string + some of the other fields that havent made it yet on GeneratinoNode
(5) Do a final pass of the generationStrategy/GenerationNode files to see what else can be migrated/condensed

Reviewed By: lena-kashtelyan

Differential Revision: D50295579