SerializationMixin.deserialize_init_args signature to take decoder_registry and class_decoder_registry (#1939)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1939

Descendents of SerializationMixin (Metrics, Runners, and Data objects) should be able to deserialize init args that are themselves Ax types. This is enabled by passing in a decoding registries, which individual deserialization functions can optionally employ. This is necessary because importing object_from_json into a Runner can create a circular dependency.

In this diff, registries are passed in but never used, so this is a no-op.

Reviewed By: lena-kashtelyan

Differential Revision: D50665111

fbshipit-source-id: ac7f5aff70e59de438b5cc953266d53d54f318dd