Remove _check_previous_required_observation method (#1999)
Summary:

This method is redundant and checks for a state that we should never reach. In the process of updating to GenerationNode based GS, maintaining this method would entail keeping track of previous nodes. The tech debt of doing so outweighs the potential reward of keeping the function. Also I think since use_update has been deprecated this function is also less relevant.

Reviewed By: lena-kashtelyan

Differential Revision: D51420534