Change DerelativizeTransform to not use model predictions when `use_raw_status_quo` is `True` or when the status quo is infeasible. (#2036)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2036

The was previously ignoring `use_raw_status_quo` and was using the values predicted by the model unless this errored out. This is misleading since we shouldn't be using the model predictions when `use_raw_status_quo` is `True`.

This also resulted in weird behavior in the case where the status quo arm was within the search space bounds but didn't satisfy the parameter constraints. This transform would then use the model to predict the metrics of the status quo (ignoring that `use_raw_status_quo` was `True`), but the model wasn't trained on the status quo since it didn't satisfy the constraints. Thus, this resulted in the model having to extrapolate and producing very weird predictions.

This diff changes the behavior to only use the model to predict the status quo metrics when (1) `use_raw_status_quo` is `False` and (2) the status quo is actually feasible.

Reviewed By: Balandat

Differential Revision: D51690727

fbshipit-source-id: cf24e2130da6b5bc6ec27f10bc449c8eacf1ca8c