Add `extract_pending_observations` function that auto-deploys to the correct pending-points function for the use case (#2039)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2039

For automated extraction of pending points, it would be ideal to not have to separately call one of two functions: get_pending... and get_pending_..._based_on_trial_status. I think we only want to go the latter route when: all trials are 1-arm and there are many trials.
Planning to use this for pending points extraction from within the ExternalGenerationStrategy (in time we might use this in other places, too, e.g. the Scheduler).

Differential Revision: D51684966

fbshipit-source-id: 4bc2ee0f8dfe69801cbc1da9d16bb8f3054390be