Support GenerationNodes and GenerationSteps in GenerationStrategy and default to GenerationNodes (#2024)
Summary:

This diff does the following:
Supports GenerationNodes at the level of GenerationStrategy. This is the big hurrah diff! 

upcoming:
(0) Add decorator for functions that are only supported in steps
(1) update the storage to include nodes independently (and not just as part of step)
(2) delete now unused GenStep functions
(3) final pass on all the doc strings and variables -- lots to clean up here
(4) add transition criterion to the repr string + some of the other fields that havent made it yet on GeneratinoNode
(5) Do a final pass of the generationStrategy/GenerationNode files to see what else can be migrated/condensed
(6) rename transiton criterion to action criterion
(7) remove conditionals for legacy usecase
(8) clean up any lingering todos

Reviewed By: lena-kashtelyan

Differential Revision: D51120002