Allowing for treating control mean as constant in relativize and unrelativize
Summary:
The `unrelativize` transform doesn't always work correctly when the inputs (particulary `means_t` and `sems_t`) are not having the proper value inside the delta method's codomain, we may observe negative `s_t2` value, the unrelativized variance.

Specifically, in `unrelativize` if we wish to perform the reverse operation of delta method, when `sems_t > sem_c * mean_c * (means_t + 1)` is not satisfied, we will see the unrelativized treatment variance to be negative, causing errors.

This diff bypasses the problem by additing an optional argument `control_as_constant` to `relativize` and `unrelativize`, which are used in the `Relativize` Ax transform for modeling. In visualization, `relativize` with `control_as_constant=False` is still being called and should still work the same.

Differential Revision: D51078296