Modelbridge outcome order matches order in datasets (#2096)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2096

Currently the torch modelbridge specifies an order for outcomes (`TorchModelBridge.outcomes`) and expects the Model to return a tensor in which outcomes follow that same order.

When each outcome has a separate model, this is done by simply ordering those models in the model list according to `outcomes`.

Things become more complicated when ContextualDatasets are involved, and now a single model might be responsible for several metrics. Currently for contextual multi-output models, this is handled via the MixedOutputModelListGP being able to re-order the outputs of the models in the list to match `outcomes`.

This really is backwards; it should be the responsibility of the ModelBridge to order things in a convenient way, and not the responsibility of the Model to adapt to whatever arbitrary order the modelbridge comes up with (alphabetically sorted, currently).

Now, I've made it the behavior of the modelbridge for `outcomes` to have the same order as the outcomes in the Datasets that are sent over to the Model at fit time. Returning things in the same order as the Dataset outcomes is a much more natural expectation on the Model.

Reviewed By: saitcakmak

Differential Revision: D51906856

fbshipit-source-id: 338b4bbc843f22cea453e4801c79f9a615192298