Default `NaN` value for model fit metrics (#2155)
Summary:

This diff sets the default value for model fit metrics to NaN in Ax telemetry. This change is made in the `scheduler.py` and `optimization.py` files. The `telemetry/tests/test_scheduler.py` and `telemetry/tests/test_optimization.py` files are also updated to reflect this change. The `utils/testing/utils.py` file is also updated to include a new function `assert_dicts_are_almost_equal` which compares two dictionaries and checks if they are almost equal. This function is used in the `telemetry/tests/test_scheduler.py` and `telemetry/tests/test_optimization.py` files to compare the expected and actual dictionaries.

Reviewed By: bernardbeckerman, Balandat

Differential Revision: D53047028