Fetch data for completed trials where appropriate (#2152)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2152

For certain types of metrics, we want to fetch data even after it is completed.

This diff also eliminates the notion of automatically using `overwrite_existing_data` for a running trial, and users can pass the scheduler option `fetcher_kwargs={"overwrite_existing_data": True}` if needed.  I don't think it's necessary because the same metrics are being fetched whether the trial is completed, and that was the concern in https://www.internalfb.com/diff/D33278531?dst_version_fbid=459585549008802&transaction_fbid=3108181146061663

Reviewed By: sdaulton

Differential Revision: D52979197

fbshipit-source-id: 17f3986a0fef2f184fbade2ed63660fb0680e762