Added missing space in error message (#2162)
Summary:
In the error message "Raw data for an arm is expected to either be a tuple of numerical mean and SEM or just a numerical mean.Got: None for metric 'result'." there was a missing space between "mean." and "Got". I've just added that single space.

Pull Request resolved: https://github.com/facebook/Ax/pull/2162

Reviewed By: bernardbeckerman

Differential Revision: D53226662

Pulled By: Balandat

fbshipit-source-id: 0153e993cd96aaafb326c0408c1d54942a48ff09