Fix GenerationStrategy and GenerataionNode todos (#2142)
Summary:

X-link: https://github.com/facebookresearch/aepsych/pull/330

In this diff we update the Ax GenerationStrategy code to remove all todos related to aepsych, and use the standard GS code flow for aepsych usecases. This required some minimal updates to the aepsych criterion and one of the storage tests.

In following diffs we will:
- revisit storage
- update AEPsych GSs as needed
- determine if run indefinetly can be replaced by simply having gen_unlimited_trials = true

Reviewed By: lena-kashtelyan

Differential Revision: D52898268