Simplify one-hot transform & support heterogeneous search spaces
Summary:
Prior to this diff, OneHot transform would construct the encoders based on the search space used while initializing the transform and apply the same encoding in `transform_search_space` regardless what parameter values the new search space included. This would make it possible to generate candidates that were not valid parameterizations of the target search space (if a larger search space was used to initialize the transform).

In addition `OneHotEncoder` was obscuring the internals of such a simple transform between two sklearn modules. These modules did not keep original ordering of the parameter values, making it unnecessarily difficult to debug & support heteregeneous search spaces. While at it, I rewrote `OneHotEncoder` using native Python operations, which simplified the code and made it easier to debug.

Differential Revision: D53449566