Simplify one-hot transform & support heterogeneous search spaces (#2181)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2181

Prior to this diff, OneHot transform would construct the encoders based on the search space used while initializing the transform and apply the same encoding in `transform_search_space` regardless what parameter values the new search space included. This would make it possible to generate candidates that were not valid parameterizations of the target search space (if a larger search space was used to initialize the transform).

In addition `OneHotEncoder` was obscuring the internals of such a simple transform between two sklearn modules. These modules did not keep original ordering of the parameter values, making it unnecessarily difficult to debug & support heteregeneous search spaces. While at it, I rewrote `OneHotEncoder` using native Python operations, which simplified the code and made it easier to debug.

Reviewed By: susanxia1006

Differential Revision: D53449566

fbshipit-source-id: 90bdd61398c0331af01b484fffe078a408df97cd