Update context order handling in ContextualDataset
Summary: Currently ContextualDataset will re-order outputs to match the specified order of context_buckets. This changes that behaviors so that the output order is always the order that the outcomes are provided in the datasets. If the user wants a particular output order, that can be handled by ordering the input datasets accordingly rather than by specifying context buckets. With this change, the context_buckets input is no longer necessary since the actual bucket names are the keys of metric_decomposition.

Differential Revision: D52452236