Multi-surrogate predictions and output ordering in MBM (#2205)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2205

MBM supports multiple surrogates, but currently doesn't support making predictions with multiple surrogates. This adds that functionality.

To do so, we have to consider how outcomes should be ordered. Consistent with D51906856 and the diff this is stacked on (D52452236), the outputs should be ordered in the same way as the input datasets. This is what will happen by default with a single surrogate. With multiple surrogates, we have to do some reordering because adjacent datasets in the input datasets list may not be modeled by the same surrogate. So this adds logic for doing that.

As part of properly handling outcomes with multiple surrogates, this diff removes the "metric_name" argument to Surrogate. The strategy of "return outputs in the same order as inputs" makes it unnecessary, since the outcome names are already part of the Datasets that are the inputs to Surrogate. Having the outcome names also specified separately is confusing and likely to lead to bugs in the future.

Because I was doing a lot of refactoring around outcome names, whenever it involved functions using the term "metric_name", I renamed that to "outcome_name".

Reviewed By: saitcakmak

Differential Revision: D52452234