`is_eligible` for multi-objective early stopping decisions (#2202)
Summary:

This commit changes the dataframe `df` argument of `is_eligible` for early stopping strategies to a `metric_dict` dictionary, mapping metric names to the dataframe containing the data corresponding to the metric.

This is in preparation for multi-objective early stopping strategies that require that data for all metrics pass the eligibility check.

While technically breaking, it appears `is_eligible` is mainly called from `should_stop_trial_early`, which itself is mainly called from `should_stop_trials_early`, so the latter appears to be the main public entry point and is not affected by the API change here.

Reviewed By: Balandat

Differential Revision: D53827530