Updating from "pivot" to "pivot_table" to fix errors generating Pareto Frontier charts (#2213)
Summary:



In the logs of a fbleaner run, the creation of the pareto frontier charts fails in “best_point.py”. The process errors out with the following exception: 
```
ValueError: Length of passed values is 1248, index implies 2.
```

When investigating, I found that this "index implies 2" error is
- Caused by Pandas pivot
- Only occurs in Panda versions prior to 1.1.12
- Happens sporadically on the same set of data

Through investigation I found that FbLearner runs with Pandas version 1.0.4, and Bento runs with Pandas version 2.0.3. This explains why the issue occurs only in FbLearner, and only occurs for some FBLearner runs. 

The fix is to use "pivot_table" instead which gives the desired output without introducing this error

Reviewed By: Balandat

Differential Revision: D54006659