Fix width issue with parameter sensitivity plot (#2215)
Summary:

In AxSweep, with high frequency the parameter sensitivity plot is useless because it has some fixed default width that is too small when the parameters have long names. Here are 3 examples of this happening in just the last week: f533780224, f532360510, f533442344

This diff fixes the issue by specify a width that grows with the size of either the parameter labels or the metric names in the dropdown.

Reviewed By: sdaulton

Differential Revision: D54021262