Allow "sort_values" to be specified in make_experiment
Summary:
When making ChoiceParameters, we see 
```
UserWarning: sort_values is not specified for ChoiceParameter "xxx". Defaulting to True for parameters of ParameterType FLOAT. To override this behavior (or avoid this warning), specify sort_values during ChoiceParameter construction.
```

However, when making Choice parameters using the ```parameter_from_json``` interface, there was currently no way to pass in the "sort_values" flag. This change adds that functionality

Reviewed By: bernardbeckerman

Differential Revision: D54315481