remove restriction on unique start_time, end_time in Data
Summary:
This removes the restriction on that `start_time` and `end_time` be the same across all rows in `Data` for those columns to be retained.

We need to retain these in order to leverage them in the model.

Removing this broke a couple of tests that verify that the `Modelbridge` has a `status_quo`---the `status_quo` was no longer set because there were multiple status_quo observations. Therefore, I added a default that when status_quo features are not specified and there are multiple status_quo observations, we should use the first observation (arbitrary) as the status quo on the `Modelbridge`. This enables using `Modelbridge.status_quo_by_trial` which in used in various `Transform`s. We still raise the warning, that really status_quo_features shoudl be used.

Reviewed By: mgarrard

Differential Revision: D54017025