Handle errors in score trace computations (#2263)
Summary:

This part of the code errors out if benchmarks are ran with a non-tranditional GS, such as one that uses GNodes rather than GSteps. Simply handling the error and skipping the score computation to unblock ad-hoc benchmarking in the short term. We can re-visit this once we start using generation nodes in nightly benchmarks.

Reviewed By: esantorella

Differential Revision: D54702411