ExternalGenerationNode (#2266)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2266

Implements a `ExternalGenerationNode` class, which defers to arbitrary non-Ax / BoTorch based methods for candidate generation. This class mostly retains the signature of the `GenerationNode` for compatibility with `GenerationStrategy`, and requires the user to implement only the necessary bits. Since it is compatible with `GenerationStrategy`, it can be easily combined with other `GenerationNode`, e.g., with Sobol to use a shared initialization with other generation strategies. This makes it particularly suitable for benchmarking against other methods from Ax.

The user needs to implement `__init__`, `update_model_state` and `get_next_trial_parameters`.

Reviewed By: Balandat

Differential Revision: D54500745

fbshipit-source-id: 5dd2dafb74c1b4089b9bfeb54338044783533d7b