Fix flaky test for sensitivity analysis
Summary:
This fixes a flaky test for sensitivity analysis. Currently the QMC
integration done for computing Sobol indices is seeded for reproducibility.
That done for computing derivative-based sensitivity measures is not. This diff
changes that so that the derivative-based measures will also have seeded QMC
with the same seed, to match Sobol indices.

A side effect of this is that it fixes the flaky test, since the results are no
longer random based on the QMC points.

Reviewed By: bernardbeckerman

Differential Revision: D54855136

fbshipit-source-id: f438de77ba85c334431b3e6c27cd8c76911a6196