Reap deprecated argument `transform_outcomes_and_configs` in `get_pareto_frontier_and_configs`
Summary: This was deprecated in v0.2.9 in  https://github.com/facebook/Ax/pull/1200. On the usual deprecation schedule, it should only be removed on 4.0.0, while we're only at 0.3.7. However, I think this is fine to remove now since the change is almost a year and a half old. `get_pareto_frontier_and_configs` is spitting out an annoying warning about the default behavior changing even when this parameter is not passed; that could in theory be removed on its own, with the argument reaped later, but I don't think it justifies a second PR.

Reviewed By: lena-kashtelyan

Differential Revision: D54902341